% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-inflation.R
\docType{data}
\name{CPI}
\alias{CPI}
\title{Consumer Price Inflation (CPI) Medicinal Products}
\format{data.frame with a row for each year. Variables are:
\describe{
   \item{year}{to year}
   \item{rate}{Proportion change from previous year}
}}
\source{
\url{https://www.ons.gov.uk/economy/inflationandpriceindices/datasets/consumerpriceinflation}
}
\usage{
CPI
}
\description{
Consumer Price Inflation (CPI) Medicinal Products
}
\keyword{datasets}
