% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflation_adjust_cost_custom.R
\name{inflation_adjust_cost_custom}
\alias{inflation_adjust_cost_custom}
\title{Inflation adjust a cost to present value}
\usage{
inflation_adjust_cost_custom(from_year, to_year, from_cost, inflation_df)
}
\arguments{
\item{from_year}{base year to inflate from}

\item{to_year}{destination year to inflate to}

\item{from_cost}{base year cost}

\item{inflation_df}{inflation annual rate data}
}
\description{
Inflation adjust a cost to present value
}
\examples{
inflation_df <- data.frame("year" = 2008:2010, "rate"=c(0.02, 0.03, 0.025))
inflation_adjust_cost_custom(2008,2009,100,inflation_df)
Result: 100*1.02 = 102 

}
