% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflation_adjust_costs_data.R
\name{inflation_adjust_cost_data}
\alias{inflation_adjust_cost_data}
\title{Calculate annual inflation adjusted costs}
\usage{
inflation_adjust_cost_data(from_year, to_year, from_cost, inflation_df_nm)
}
\arguments{
\item{from_year}{Date of cost to convert from}

\item{to_year}{Date to convert cost to}

\item{from_cost}{Cost at \code{from_year}}

\item{inflation_df_nm}{Source of data (string) \code{\link{HCHS}}, or \code{\link{CPI}}.}
}
\value{
Inflated (to) cost (scalar), with attributes used to generate the return value:
\itemize{
  \item \code{from_year}
  \item \code{to_year}
  \item \code{from_cost}
  \item \code{reference} Which inflation adjustment dataset used.
 }
}
\description{
Up to a present value (PV) inflated upwards.
}
\details{
Option to use the following datasets:

\itemize{
  \item \href{https://www.pssru.ac.uk/pub/uc/uc2017/sources-of-information.pdf}{PSSRU annual inflation hospital and community health services (HCHS)}
  \item \link{CPI}
}

\deqn{(1 + i_{1})(1 + i_{2}) \cdots (1 + i_{n}) \times C}

Clearly, for the same \code{i} each year this is simply
\deqn{C(1 + i)^n}
}
\examples{
from_year <- 2012
to_year <- 2015
from_cost <- 96.140

1*(1+0.035)^6

inflation_adjust_cost_data(from_year = 2014,
                           to_year = 2016,
                           from_cost = 1,
                           inflation_df_nm = "HCHS")
1*1.004*1.01

}
